+++ src/main/java/io/github/mxx1111/medagentguard/fhir/FhirBundleValidationService.java
@@ -1,0 +1,6 @@
+package io.github.mxx1111.medagentguard.fhir;
+
+public interface FhirBundleValidationService {
+
+    FhirBundleValidationResult validate(String fhirJson);
+}